#13 refactor: compose layer rename into presentation
