--- Ppap/app/src/androidTest/java/com/jeongg/ppap/compose/SubscribeComposeTest.kt
@@ -1,89 +1,0 @@
-package com.jeongg.ppap.compose
-
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.test.onNodeWithContentDescription
-import androidx.compose.ui.test.performClick
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import com.jeongg.ppap.compose.subscribe.CustomSubscribe
-import com.jeongg.ppap.compose.subscribe.CustomSubscribeItem
-import com.jeongg.ppap.compose.subscribe.DefaultSubscribeItem
-import org.junit.Assert.*
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class SubscribeComposeTest {
-    @get: Rule
-    val composeTestRule = createComposeRule()
-
-    private val checked = composeTestRule.onNodeWithContentDescription("checked: true")
-    private val unchecked = composeTestRule.onNodeWithContentDescription("checked: false")
-    private val emptyContent = composeTestRule.onNode(hasTestTag("empty_content"))
-
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.jeongg.ppap", appContext.packageName)
-    }
-    @Test
-    fun subscribe_default_checkedToggle(){
-        val isSelected = true
-        start_default_subscribe_item(isSelected)
-
-        checked.assertIsDisplayed()
-        unchecked.assertDoesNotExist()
-
-        checked.performClick() // checked -> unchecked
-
-        unchecked.assertIsDisplayed()
-        checked.assertDoesNotExist()
-    }
-    @Test
-    fun subscribe_custom_uncheckedToggle(){
-        val isSelected = false
-        start_custom_subscribe_item(isSelected)
-
-        unchecked.assertIsDisplayed()
-        checked.assertDoesNotExist()
-
-        unchecked.performClick() // checked -> unchecked
-
-        checked.assertIsDisplayed()
-        unchecked.assertDoesNotExist()
-    }
-    @Test
-    fun subscribe_emptyItem(){
-        val subscribes = emptyList<String>()
-        start_custom_subscribe(subscribes)
-
-        emptyContent.assertIsDisplayed()
-    }
-    @Test
-    fun subscribe_nonEmptyItem(){
-        val subcribes = listOf("최강 정컴", "전기 최고", "얍")
-        start_custom_subscribe(subcribes)
-
-        emptyContent.assertDoesNotExist()
-    }
-
-    private fun start_default_subscribe_item(isChecked: Boolean = false){
-        composeTestRule.setContent {
-            DefaultSubscribeItem(isSelected = isChecked)
-        }
-    }
-    private fun start_custom_subscribe_item(isChecked: Boolean = false){
-        composeTestRule.setContent {
-            CustomSubscribeItem(isSelected = isChecked)
-        }
-    }
-    private fun start_custom_subscribe(subscribeList: List<String> = emptyList()){
-        composeTestRule.setContent {
-            CustomSubscribe(subscribes = subscribeList)
-        }
-    }
-}
