--- Ppap/app/src/main/java/com/jeongg/ppap/compose/theme/Shape.kt
@@ -1,12 +1,0 @@
-package com.jeongg.ppap.compose.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Shapes
-import androidx.compose.ui.unit.dp
-
-val shapes = Shapes(
-    extraLarge = RoundedCornerShape(45.dp),
-    large = RoundedCornerShape(20.dp),
-    medium = RoundedCornerShape(15.dp),
-    small = RoundedCornerShape(10.dp)
-)
