+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/component/PDivider.kt
@@ -1,0 +1,19 @@
+package com.jeongg.ppap.presentation.component
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Divider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PDivider(
+    modifier: Modifier = Modifier
+){
+    Divider(
+        modifier = modifier.fillMaxWidth(),
+        thickness = 1.dp,
+        color = MaterialTheme.colorScheme.outline
+    )
+}
