+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/theme/Color.kt
@@ -1,0 +1,27 @@
+package com.jeongg.ppap.presentation.theme
+
+import androidx.compose.ui.graphics.Color
+
+val very_bright_yellow = Color(0xFFFDE27F)
+val bright_yellow = Color(0xFFFED84D)
+val main_yellow = Color(0xFFFDC700)
+
+val main_green = Color(0xFFBBD699)
+
+val bright_pink = Color(0xFFFCEDED)
+val main_pink = Color(0xFFF7D0D1)
+
+val gray1 = Color(0xFFEBEBEB)
+val gray2 = Color(0xFF808080)
+val gray3 = Color(0xFF979797)
+val gray4 = Color(0xFF656565)
+
+val rainbowColorsBrush =
+    listOf(
+        Color(0xFFE57373),
+        Color(0xFFFFB74D),
+        Color(0xFFFFF176),
+        Color(0xFFAED581),
+        Color(0xFF4DD0E1),
+        Color(0xFF9575CD)
+    )
