+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/theme/Shape.kt
@@ -1,0 +1,12 @@
+package com.jeongg.ppap.presentation.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Shapes
+import androidx.compose.ui.unit.dp
+
+val shapes = Shapes(
+    extraLarge = RoundedCornerShape(45.dp),
+    large = RoundedCornerShape(20.dp),
+    medium = RoundedCornerShape(15.dp),
+    small = RoundedCornerShape(10.dp)
+)
