+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/theme/Theme.kt
@@ -1,0 +1,45 @@
+package com.jeongg.ppap.presentation.theme
+
+import android.os.Build
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+
+private val DarkColorScheme = darkColorScheme(
+    background = Color.Black,
+    surface = Color.White,
+    outline = gray3,
+    outlineVariant = gray3
+)
+
+private val LightColorScheme = lightColorScheme(
+    background = Color.White,
+    surface = Color.Black,
+    outline = gray1,
+    outlineVariant = gray4
+)
+
+@Composable
+fun PpapTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    dynamicColor: Boolean = true,
+    content: @Composable () -> Unit
+) {
+    val colorScheme = when {
+        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+            if (darkTheme) DarkColorScheme else LightColorScheme
+        }
+        darkTheme -> DarkColorScheme
+        else -> LightColorScheme
+    }
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        typography = typography,
+        shapes = shapes,
+        content = content
+    )
+}
