+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/theme/Type.kt
@@ -1,0 +1,72 @@
+package com.jeongg.ppap.presentation.theme
+
+import androidx.compose.material3.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import com.jeongg.ppap.R
+
+// Set of Material typography styles to start with
+val RobotoFamily = FontFamily(
+    Font(R.font.bold, FontWeight.Bold, FontStyle.Normal),
+    Font(R.font.extra_bold, FontWeight.ExtraBold, FontStyle.Normal),
+    Font(R.font.extra_light, FontWeight.ExtraLight, FontStyle.Normal),
+    Font(R.font.light, FontWeight.Light, FontStyle.Normal),
+    Font(R.font.medium, FontWeight.Medium, FontStyle.Normal),
+    Font(R.font.pretendard_black, FontWeight.Black, FontStyle.Normal),
+    Font(R.font.regular, FontWeight.Normal, FontStyle.Normal),
+    Font(R.font.semi_bold, FontWeight.SemiBold, FontStyle.Normal),
+    Font(R.font.thin, FontWeight.Thin, FontStyle.Normal),
+)
+
+val typography = Typography(
+    headlineLarge = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Bold,
+        fontSize = 50.sp,
+    ),
+    titleLarge = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Bold,
+        fontSize = 20.sp,
+    ),
+    titleMedium = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Bold,
+        fontSize = 18.sp,
+    ),
+    titleSmall = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Bold,
+        fontSize = 15.sp,
+    ),
+    displayLarge = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.SemiBold,
+        fontSize = 16.sp,
+    ),
+    bodyLarge = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Medium,
+        fontSize = 15.sp,
+        lineHeight = 20.sp
+    ),
+    bodyMedium = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Medium,
+        fontSize = 13.sp,
+    ),
+    labelLarge = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Normal,
+        fontSize = 12.sp,
+    ),
+    labelMedium = TextStyle(
+        fontFamily = RobotoFamily,
+        fontWeight = FontWeight.Normal,
+        fontSize = 10.sp,
+    )
+)
